Simplify boiler plate code for everything
The previous iteration of this was a bit unergonomic. It changes a few
things under the hood, but it adds two convenience macros `router` and
`routes`.

`router` is used to assign handlers to specific routes.
`routes` assigns, Method verbs, handler name, and a function to deal
with requests to a type that gets passed into router.

This makes it nicer to deal with as an end user, but still provides the
flexibility and low overhead for the framework.