Add ability to have a CORS Preflight hanlder.
The macro for `router!()` has been changed to take a only a Route
name like (Cors) at the end. This will implicitly add it for all routes.

However, you need to still specify it's method as Options in order for
it to work the way you want. This is not middleware and won't work that
way. You'll need to write the handler to send back the Cors request you
want to send back.